BOARD_MEGA: Correctly hook xmem into the avr-libc malloc freelist. For all boards, free memory is more correctly gauged by taking the freelist into account.

+++ Memory.h
@@ -5,7 +5,7 @@ class MemoryController
 {
 public:
     void scanForRAM (void);
-    uint16_t available (void);
+    int available (void);
 };
 
 extern MemoryController Memory;
